fix files

+++ src/category/category.service.ts
@@ -141,7 +141,9 @@ export class CategoryService {
 				id: dto.id
 			},
 		})
-		if(files){
+
+		if(files.mainImg && Array.isArray(files.mainImg) && files.mainImg.length > 0){
+			console.log(files)
 			const mainImgFile = files.mainImg ? files.mainImg[0] : null;
 			const iconImgFile = files.iconImg ? files.iconImg[0] : null;
 			deleteFile(category.iconImg)
@@ -185,7 +187,7 @@ export class CategoryService {
 			id: dto.id,
 		}
 	})		
-		if(files){
+		if(files.mainImg && Array.isArray(files.mainImg) && files.mainImg.length > 0){
 			
 	
 	if(!subcategory){
